requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
datapizza-ai==0.0.2
datapizza-ai-clients-google==0.0.3
datapizza-ai-clients-openai==0.0.5
datapizza-ai-clients-openai-like==0.0.5
datapizza-ai-core==0.0.5
datapizza-ai-embedders-openai==0.0.4
datapizza-ai-vectorstores-qdrant==0.0.2
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-genai==1.45.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.9.0
numpy==2.3.4
openai==2.6.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
zipp==3.23.0
